allow xnat authentication data from ~/.xnat_auth, command line arguments, or environment variables; switch over to StoreXAR command line utility; add --keep-xar; add Dockerfile

+++ anatqc/cli/process.py
@@ -16,6 +16,7 @@ from anatqc.xnat import Report
 import anatqc.tasks.mriqc as mriqc
 import anatqc.tasks.vnav as vnav
 import anatqc.tasks.morph as morph
+import anatqc.cli.get
 from anatqc.state import State
 
 logger = logging.getLogger(__name__)
@@ -105,7 +106,8 @@ def do(args):
             logger.info('building xnat archive file %s', fo.name)
             R.build_assessment(fo)
         logger.info('uploading %s to %s', fo.name, args.xnat_upload)
-        auth = yaxil.auth(args.xnat_upload)
-        yaxil.storexar(auth, fo.name)
-        logger.info('removing %s', fo.name)
-        os.remove(fo.name)
+        auth = anatqc.cli.get.getauth(args)
+        yaxil.storexar_cli(auth, fo.name)
+        if not args.keep_xar:
+            logger.info('removing %s', fo.name)
+            os.remove(fo.name)
